requirements.txt
Flask==2.3.3
deepface==0.0.87
opencv-python==4.8.1.78
tensorflow==2.13.0
Pillow==10.0.1
numpy==1.24.3